hash function test
